build(): set clean architecture
