+++ app/src/main/java/fr/idnow/imagecapture/data/dto/QuoteDTO.kt
@@ -1,0 +1,4 @@
+package fr.idnow.imagecapture.data.dto
+
+class QuoteDTO {
+}
