+++ app/src/main/java/fr/idnow/imagecapture/data/repositories/RemoteQuoteRepository.kt
@@ -1,0 +1,4 @@
+package fr.idnow.imagecapture.data.repositories
+
+class RemoteQuoteRepository {
+}
