+++ app/src/main/java/fr/idnow/imagecapture/domain/entities/Quote.kt
@@ -1,0 +1,3 @@
+package fr.idnow.imagecapture.domain.entities
+
+data class Quote(val text: String)
