+++ app/src/main/java/fr/idnow/imagecapture/domain/repositories/QuoteRepository.kt
@@ -1,0 +1,4 @@
+package fr.idnow.imagecapture.domain.repositories
+
+interface QuoteRepository {
+}
