+++ app/src/main/java/fr/idnow/imagecapture/domain/usecases/GetSingleQuoteUseCase.kt
@@ -1,0 +1,4 @@
+package fr.idnow.imagecapture.domain.usecases
+
+class GetSingleQuoteUseCase {
+}
