+++ app/src/main/java/fr/idnow/imagecapture/presentation/viewmodels/QuoteViewModel.kt
@@ -1,0 +1,4 @@
+package fr.idnow.imagecapture.presentation.viewmodels
+
+class QuoteViewModel {
+}
